requirements.txt
flask
python-dotenv
networkx
openai
